requirements.txt
yfinance
pandas
stockstats
requests
python-dotenv
